Best Model
This is the best model (an SVM with a linear kernel) trained on the data with a 5-fold cross validation approach. The features are first sintetized with the PCA techniques; after that, the model consider only the first two components that explain more or less the 90% of the total variability.